% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cell.cycle.obj}
\alias{cell.cycle.obj}
\title{Cell cycling signatures}
\format{
A list of cycling signatures.
}
\source{
\doi{10.1126/science.aad0501}
}
\usage{
cell.cycle.obj
}
\description{
A list of cell cycling signatures (G1.S and G2.M phases),
for mouse and human.
}
\keyword{datasets}
