% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cellstate.predict}
\alias{cellstate.predict}
\title{Predict cell states of a projected dataset}
\usage{
cellstate.predict(
  ref,
  query,
  reduction = "pca",
  ndim = 10,
  k = 20,
  labels.col = "functional.cluster"
)
}
\arguments{
\item{ref}{Reference Atlas}

\item{query}{Seurat object with query data}

\item{reduction}{The dimensionality reduction used to calculate pairwise distances. One of "pca" or "umap"}

\item{ndim}{How many dimensions in the reduced space to be used for distance calculations}

\item{k}{Number of neighbors to assign the cell type}

\item{labels.col}{The metadata field of the reference to annotate the clusters (default: functional.cluster)}
}
\value{
The query object submitted as parameter, with two additional metadata slots for predicted state and its confidence score
}
\description{
This function uses a nearest-neighbor algorithm to predict a feature (e.g. the cell state) of the query cells. Distances between
cells in the reference map and cells in the query are calculated in a reduced space (PCA or UMAP) and the feature is assigned to
query cells based on a consensus of its nearest neighbors in the reference object.
}
\examples{
cellstate.predict(ref, query_example.seurat)
}
