% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute_silhouette}
\alias{compute_silhouette}
\title{Calculate Silhouette coefficient}
\usage{
compute_silhouette(
  ref,
  query = NULL,
  reduction = "pca",
  ndim = NULL,
  label_col = "functional.cluster",
  normalize.scores = FALSE,
  min.cells = 20
)
}
\arguments{
\item{ref}{Reference object}

\item{query}{Query object. If not specified, the silhouette coefficient of only the reference will be calculated}

\item{reduction}{Which dimensionality reduction to use for euclidian distance calculation}

\item{ndim}{Number of dimensions in the dimred to use for distance calculation. If NULL, use all dimensions.}

\item{label_col}{Metadata column with cell type annotations. Must be present both in reference and query}

\item{normalize.scores}{Whether to normalize silhouette scores by the average cell type silhouettes of the reference}

\item{min.cells}{Only report silhouette scores for cell type with at least this number of cells}
}
\value{
A dataframe with average silhouette coefficient for each cell type
}
\description{
Given a projected object and its reference, calculate silhouette coefficient for query cells with respect
to reference cells with the same cell labels.
}
\examples{
data(query_example_seurat)
ref <- load.reference.map()
q <- Run.ProjecTILs(query_example_seurat, ref=ref, fast.umap.predict=TRUE)
combined <- compute_silhouette(ref, query=q)
}
