% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get.reference.maps}
\alias{get.reference.maps}
\title{Retrieve and load reference atlas}
\usage{
get.reference.maps(
  collection = NULL,
  reference = NULL,
  update = FALSE,
  directory = "./ProjecTILs_references",
  as.list = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{collection}{Collection to download and load. See available collection using \link{list.reference.maps}. If NULL, all are downloaded and loaded (default)}

\item{reference}{References to download and load. See available collection using \link{list.reference.maps}. If NULL, all are downloaded and loaded (default)}

\item{update}{Boolean whether to delete current reference maps and download them again}

\item{directory}{Directory where to download and load from reference maps. By default a directory named "ProjecTILs_references" is created in working directory.}

\item{as.list}{Boolean whether to simplify list (\code{FALSE}) or, by default, keep a list of lists for each collection (\code{TRUE}).}

\item{verbose}{Inform of the status of processes}
}
\description{
Download and load reference atlases.
}
\examples{
# explore available reference maps
list.reference.maps()

# consider increasing downloading timeout
options(timeout = 1000)

# get all available reference maps
ref.maps <- get.reference.maps()

# get certain collections or reference maps
# all human references maps
ref.maps.human <- get.reference.maps(collection = "human")

# only some references
ref.maps <- get.reference.maps(reference = "DC")
ref.maps.CD4 <- get.reference.maps(reference = c("CD4", "Virus_CD4T"))

# update previously downloaded maps
ref.maps <- get.reference.maps(update = TRUE)

}
