% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Hs2Mm.convert.table}
\alias{Hs2Mm.convert.table}
\title{Human-mouse ortholog conversion table}
\format{
A dataframe containing gene ortholog mapping.
}
\source{
\url{https://www.ensembl.org/Mus_musculus/Info/Index}
}
\usage{
Hs2Mm.convert.table
}
\description{
A conversion table of stable orthologs between Hs and Mm.
}
\keyword{datasets}
