% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{list.reference.maps}
\alias{list.reference.maps}
\title{Available reference atlas for ProjecTILs}
\usage{
list.reference.maps()
}
\description{
Obtain the list of available reference atlas for ProjecTILs to then download and load them using \link{get.reference.maps}.
}
\examples{
# explore available reference maps
list.reference.maps()
}
