% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load.reference.map}
\alias{load.reference.map}
\title{Load Reference Atlas}
\usage{
load.reference.map(ref = "referenceTIL")
}
\arguments{
\item{ref}{Reference Atlas Seurat object (by default downloads a mouse reference TIL atlas). Otherwise provide a custom reference atlas.}
}
\description{
Load or download the reference map for dataset projection. By the default it downloads the reference atlas of tumour-infiltrating lymphocytes (TILs).
}
\examples{
load.reference.map()
}
