% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make.reference}
\alias{make.reference}
\title{Make a ProjecTILs reference}
\usage{
make.reference(
  ref,
  assay = NULL,
  atlas.name = "custom_reference",
  annotation.column = "functional.cluster",
  recalculate.umap = FALSE,
  ndim = 20,
  dimred = "umap",
  nfeatures = 1000,
  seed = 123
)
}
\arguments{
\item{ref}{Seurat object with reference atlas}

\item{assay}{The data slot where to pull the expression data}

\item{atlas.name}{An optional name for your reference}

\item{annotation.column}{The metadata column with the cluster annotations for this atlas}

\item{recalculate.umap}{If TRUE, run the `umap` algorithm to generate embeddings. Otherwise use the embeddings stored in the `dimred` slot.}

\item{ndim}{Number of dimensions for PCA to be passed to the `umap` method}

\item{dimred}{Use the pre-calculated embeddings stored at `ref@reductions[[dimred]]`}

\item{nfeatures}{Number of variable features (only calculated if not already present)}

\item{seed}{Random seed}
}
\value{
A reference atlas compatible with ProjecTILs
}
\description{
Converts a Seurat object to a ProjecTILs reference atlas. You can preserve your low-dimensionality embeddings (e.g. UMAP) in the reference atlas by
setting `recalculate.umap=FALSE`, or let the method apply the `umap` package.
}
\examples{
custom_reference <- ProjecTILs::make.reference(myref, assay="integrated")  
#Add a color palette for your atlas
custom_reference@misc$atlas.palette <- c("#edbe2a","#A58AFF","#53B400","#F8766D","#00B6EB","#d1cfcc","#FF0000","#87f6a5","#e812dd")
}
