% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{merge.Seurat.embeddings}
\alias{merge.Seurat.embeddings}
\title{Merge Seurat objects, including reductions (e.g. PCA, UMAP, ICA)}
\usage{
\method{merge}{Seurat.embeddings}(x = NULL, y = NULL, merge.dr = TRUE, ...)
}
\arguments{
\item{x}{First object to merge}

\item{y}{Second object to merge}

\item{merge.dr}{How to handle merging dimensional reductions (see merge.Seurat)}

\item{...}{More parameters to \link{merge} function}
}
\value{
A merged Seurat object
}
\description{
Given two Seurat objects, merge counts and data as well as dim reductions (PCA, UMAP, ICA, etc.)
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
o1 <- query_example_seurat
o2 <- query_example_seurat
seurat.merged <- merge.Seurat.embeddings(o1, o2)
#To merge multiple object stored in a list
seurat.merged <- Reduce(f=merge.Seurat.embeddings, x=obj.list)
\dontshow{\}) # examplesIf}
}
