% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.discriminant.3d}
\alias{plot.discriminant.3d}
\title{3D plot of reference map with extra discriminant dimension}
\usage{
\method{plot}{discriminant.3d}(
  ref,
  query,
  query.control = NULL,
  query.assay = "RNA",
  labels.col = "functional.cluster",
  extra.dim = "ICA_1",
  query.state = NULL
)
}
\arguments{
\item{ref}{Seurat object with reference atlas}

\item{query}{Seurat object with query data}

\item{query.control}{Optionally, you can compare your query with a control sample, instead of the reference}

\item{query.assay}{The data slot to be used for enrichment analysis}

\item{labels.col}{The metadata field used to annotate the clusters}

\item{extra.dim}{The additional dimension to be added on the z-axis of the plot. Can be either:
\itemize{
  \item{An ICA or PCA dimension (e.g. ICA_10). See `find.discriminant.dimensions`}
  \item{Any numeric metadata field associated to the cells (e.g. 'cycling.score')}
}}

\item{query.state}{Only plot the query cells from this specific state}
}
\value{
A three dimensional plot with UMAP_1 and UMAP_2 on the x and y axis respectively, and the specified `extra.dim` on the z-axis.
}
\description{
Add an extra dimension to the reference map (it can be suggested by `find.discriminant.dimensions`), to explore additional axes of variability
in a query dataset compared to the reference map.
}
\examples{
plot.discriminant.3d(ref, query=query, extra.dim="ICA_19")
plot.discriminant.3d(ref, query=treated.set, query.control=control.set, extra.dim="ICA_2")
}
