% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.statepred.composition}
\alias{plot.statepred.composition}
\title{Summarize the predicted cell states of an object}
\usage{
\method{plot}{statepred.composition}(
  ref,
  query,
  labels.col = "functional.cluster",
  cols = NULL,
  metric = c("Count", "Percent")
)
}
\arguments{
\item{ref}{Reference object}

\item{query}{Seurat object with query data}

\item{labels.col}{The metadata field used to annotate the clusters (default: functional.cluster)}

\item{cols}{Custom color palette for clusters}

\item{metric}{One of `Count` or `Percent`. `Count` plots the absolute number of cells, `Percent` the fraction on the total number of cells.}
}
\value{
Barplot of predicted state composition
}
\description{
Makes a barplot of the frequency of cell states in a query object.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
data(query_example_seurat)
ref <- load.reference.map()
q <- make.projection(query_example_seurat, ref=ref)
q <- cellstate.predict(ref, query=q)
plot.statepred.composition(query_example.seurat)
\dontshow{\}) # examplesIf}
}
