% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.states.radar}
\alias{plot.states.radar}
\title{Show expression level of key genes}
\usage{
\method{plot}{states.radar}(
  ref,
  query = NULL,
  labels.col = "functional.cluster",
  genes4radar = NULL,
  min.cells = 10,
  cols = NULL,
  return = F,
  return.as.list = F
)
}
\arguments{
\item{ref}{Seurat object with reference atlas}

\item{query}{Query data, either as a Seurat object or as a list of Seurat objects}

\item{labels.col}{The metadata field used to annotate the clusters (default: functional.cluster)}

\item{genes4radar}{Which genes to use for plotting (default: c("Foxp3","Cd4","Cd8a","Tcf7","Ccr7","Gzmb","Gzmk","Pdcd1","Havcr2","Tox,"Mki67")}

\item{min.cells}{Only display cell states with a minimum number of cells}

\item{cols}{Custom color palette for samples in radar plot}

\item{return}{Return the combined grobs instead of printing it to the default device}

\item{return.as.list}{Return plots in a list, instead of combining them in a single plot}
}
\value{
Radar plot of gene expression of key genes by cell subtype
}
\description{
Makes a radar plot of the expression level of a set of genes. It can be useful to compare the gene expression profile of different cell
states in the reference atlas vs. a projected set.
}
\examples{
plot.states.radar(ref)
}
