% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{query_example_seurat}
\alias{query_example_seurat}
\title{Test dataset for ProjecTILs}
\format{
A Seurat object
}
\source{
\url{https://pmc.ncbi.nlm.nih.gov/articles/PMC6673650/}
}
\usage{
query_example_seurat
}
\description{
A small dataset of CD8 T cells, to test the ProjecTILs installation.
}
\keyword{datasets}
