% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read.sc.query}
\alias{read.sc.query}
\title{Read to memory a query expression matrix}
\usage{
read.sc.query(
  filename,
  type = c("10x", "hdf5", "raw", "raw.log2"),
  project.name = "Query",
  min.cells = 3,
  min.features = 50,
  gene.column.10x = 2,
  raw.rownames = 1,
  raw.sep = c("auto", " ", "\\t", ","),
  raw.header = TRUE,
  use.readmtx = TRUE
)
}
\arguments{
\item{filename}{Path to expression matrix file or folder}

\item{type}{Expression matrix format (10x, hdf5, raw, raw.log2)}

\item{project.name}{Title for the project}

\item{min.cells}{Only keep genes represented in at least min.cells number of cells}

\item{min.features}{Only keep cells expressing at least min.features genes}

\item{gene.column.10x}{For 10x format - which column of genes.tsv or features.tsv to use for gene names}

\item{raw.rownames}{For raw matrix format - A vector of row names, or a single number giving the column of the table which contains the row names}

\item{raw.sep}{For raw matrix format - Separator for raw expression matrix}

\item{raw.header}{For raw matrix format - Use headers in expression matrix}

\item{use.readmtx}{Use ReadMtx function to read in 10x files with custom names}
}
\value{
A Seurat object populated with raw counts and normalized counts for single-cell expression
}
\description{
Load a query expression matrix to be projected onto the reference atlas. Several formats (10x, hdf5, raw and log counts) 
are supported - see \code{type} parameter for details
}
\examples{
fname <- "./sample_data"
querydata <- read.sc.query(fname, type="10x")
}
