% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{recalculate.embeddings}
\alias{recalculate.embeddings}
\title{Recalculate low dimensional embeddings after projection}
\usage{
recalculate.embeddings(
  ref,
  projected,
  ref.assay = "integrated",
  proj.assay = "integrated",
  ndim = NULL,
  n.neighbors = 20,
  min.dist = 0.3,
  recalc.pca = FALSE,
  resol = 0.4,
  k.param = 15,
  metric = "cosine",
  umap.method = c("umap", "uwot"),
  seed = 123
)
}
\arguments{
\item{ref}{Reference map}

\item{projected}{A projected object (or list of projected objects) generated using \link{make.projection}}

\item{ref.assay}{Assay for reference object}

\item{proj.assay}{Assay for projected object(s)}

\item{ndim}{Number of dimensions for recalculating dimensionality reductions}

\item{n.neighbors}{Number of neighbors for UMAP algorithm}

\item{min.dist}{Tightness parameter for UMAP embedding}

\item{recalc.pca}{Whether to recalculate the PCA embeddings with the combined reference and projected data}

\item{resol}{Resolution for unsupervised clustering}

\item{k.param}{Number of nearest neighbors for clustering}

\item{metric}{Distance metric to use to find nearest neighbors for UMAP}

\item{umap.method}{Which method should be used to calculate UMAP embeddings}

\item{seed}{Random seed for reproducibility}
}
\value{
A combined reference object of reference and projected object(s), with new low dimensional embeddings
}
\description{
Given a reference object and a (list of) projected objects, recalculate low-dim embeddings accounting for the projected cells
}
\examples{
combined <- recalculate.embeddings(ref, projected, ndim=10)
}
