% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{celltype.heatmap}
\alias{celltype.heatmap}
\title{Plot a averaged expression heatmap from a Seurat object}
\usage{
celltype.heatmap(
  data,
  assay = "RNA",
  genes,
  ref = NULL,
  scale = "row",
  method = c("ward.D2", "ward.D", "average"),
  brewer.palette = "RdBu",
  palette_reverse = F,
  palette = NULL,
  cluster.col = "functional.cluster",
  metadata = NULL,
  order.by = NULL,
  flip = FALSE,
  cluster_genes = FALSE,
  cluster_samples = FALSE,
  min.cells = 10,
  show_samplenames = FALSE,
  remove.NA.meta = TRUE,
  breaks = seq(-2, 2, by = 0.1),
  ...
)
}
\arguments{
\item{data}{A Seurat object to be used for the heatmap}

\item{assay}{A string indicating the assay type, default is "RNA"}

\item{genes}{A vector of genes to be used in the heatmap}

\item{ref}{A ProjecTILs reference Seurat object to define the order of functional.cluster}

\item{scale}{A string indicating the scale of the heatmap, default is "row"}

\item{method}{A string or vector of strings indicating the clustering method to be used, default is "ward.D2"}

\item{brewer.palette}{A string indicating the color palette to be used, default is "RdBu"}

\item{palette_reverse}{A boolean indicating if color palette should be reversed, default is FALSE}

\item{palette}{A named list containing colors vectors compatible with pheatmap. The list is named by the metadata names, default is taking these palettes to plot metadata: "Paired","Set2","Accent","Dark2","Set1","Set3".}

\item{cluster.col}{A string indicating the column name of the functional cluster to be used}

\item{metadata}{A vector of metadata to be used in the heatmap}

\item{order.by}{A string indicating the column name to reorder the heatmap}

\item{flip}{A boolean indicating if the heatmap should be flipped, default is FALSE}

\item{cluster_genes}{A boolean indicating if genes should be clustered, default is FALSE}

\item{cluster_samples}{A boolean indicating if samples should be clustered, default is FALSE}

\item{min.cells}{A value defining the minimum number of cells a sample should have to be kept, default is 10}

\item{show_samplenames}{A boolean indicating whether the heatmap should display the sample names or not, default is FALSE}

\item{remove.NA.meta}{A boolean indicating if missing samples with missing metadata should be plotted, default is TRUE}
}
\value{
A pheatmap plot, displaying averaged expression values across genes for each selected genes and samples.
}
\description{
This function allows to plot a averaged expression heatmap starting from a Seurat object,
split by possibly any metadata present in the starting Seurat object
}
\details{
The function first loads the pheatmap package, sets a seed for reproducibility, and selects the first element of the "method" parameter as the clustering method to be used
It then selects the wanted metadata from the data set, removing any missing values if the "remove.NA.meta" parameter is set to TRUE. <- 
Next, the function filters the data set to only include samples that have at least "min.cells" in the "metaSubset" variable. It then calculates the mean expression of the selected genes by the "metaSubset" variable.
The function then reorders the data frame if the "order.by" parameter is not null, and sets up the colors for the heatmap.
Finally, the function creates the heatmap using the pheatmap package, with the "method", "scale", "flip", "cluster_genes", "cluster_samples", and "show_samplenames" parameters controlling various aspects of the heatmap.
}
\examples{
library(Seurat)
ref <- load.reference.map(ref = "https://figshare.com/ndownloader/files/38921366")
celltype.heatmap(ref, assay = "RNA", genes = c("LEF1","SELL","GZMK","FGFBP2","HAVCR2","PDCD1","XCL1","KLRB1"), ref = ref, cluster.col = "functional.cluster", metadata = c("orig.ident", "Tissue"), order.by = "Tissue")
}
