% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load.reference.map}
\alias{load.reference.map}
\title{Load Reference Atlas}
\usage{
load.reference.map(ref = "referenceTIL")
}
\arguments{
\item{ref}{Reference atlas as a Seurat object (by default downloads a mouse reference TIL atlas).
To use a custom reference atlas, provide a .rds object or a URL to a .rds object, storing a Seurat object
prepared using \link{make.reference}}
}
\description{
Load or download the reference map for dataset projection.
By the default it downloads a reference atlas of tumour-infiltrating
lymphocytes (TILs) from mouse.
}
\examples{
load.reference.map()
}
