% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make.projection}
\alias{make.projection}
\title{Project a query scRNA-seq dataset onto a reference atlas}
\usage{
make.projection(
  query,
  ref = NULL,
  filter.cells = T,
  query.assay = "auto",
  direct.projection = FALSE,
  seurat.k.filter = 200,
  skip.normalize = FALSE
)
}
\arguments{
\item{query}{Seurat object with query data}

\item{ref}{Reference Atlas Seurat object - if NULL, downloads the default TIL reference atlas}

\item{filter.cells}{Logical value indicating if T cells should be pre-filtered using TILPRED. Default is TRUE. Only set to FALSE if the dataset has been previously checked for non-T cell contaminations.}

\item{query.assay}{Which assay slot to use for the query}

\item{direct.projection}{Logical. If true, apply PCA transformation directly without alignment}

\item{seurat.k.filter}{Integer. For alignment, how many neighbors (k) to use when picking anchors. Default is 200; try lower value in case of failure}

\item{skip.normalize}{Logical. By default, log-normalize the count data. If you have already normalized your data, you can skip normalization.}
}
\value{
An augmented object \code{query} with projected UMAP coordinates on the reference map and cells classifications
}
\description{
This function allows projecting a single-cell RNA-seq dataset onto a reference map of cellular states.
It can be useful to interpret a new dataset in the context of an annotated Atlas of known cell types. See `load.reference.map` to load
or download a reference atlas.
}
\examples{
data(query_example_seurat)
make.projection(query_example_seurat)
}
