% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make.projection}
\alias{make.projection}
\title{Project a query scRNA-seq dataset onto a reference atlas}
\usage{
make.projection(
  query,
  ref = NULL,
  filter.cells = TRUE,
  query.assay = NULL,
  direct.projection = FALSE,
  batch_correction = 0.5,
  correction_scale = 0.1,
  k.anchor = 8,
  k.weight = 50,
  skip.normalize = FALSE,
  fast.mode = FALSE,
  ortholog_table = NULL,
  scGate_model = NULL,
  ncores = 1
)
}
\arguments{
\item{query}{Query data, either as single Seurat object or as a list of Seurat object}

\item{ref}{Reference Atlas - if NULL, downloads the default TIL reference atlas}

\item{filter.cells}{Pre-filter cells using `scGate`. Only set to FALSE if the dataset has been previously subset to desired cell type.}

\item{query.assay}{Which assay slot to use for the query (defaults to DefaultAssay(query))}

\item{direct.projection}{If true, apply PCA transformation directly without alignment}

\item{batch_correction}{Strength of batch-effect correction, between 0 and 1, where low values indicate weak batch correction and high values strong batch correction.}

\item{correction_scale}{Slope of sigmoid function used to determine strength of batch effect correction.}

\item{k.anchor}{Integer. For alignment, how many neighbors (k) to use when picking anchors.}

\item{k.weight}{Integer. Number of neighbors to consider when weighting anchors.}

\item{skip.normalize}{By default, log-normalize the count data. If you have already normalized your data, you can skip normalization.}

\item{fast.mode}{Fast approximation for UMAP projection. Uses coordinates of nearest neighbors in PCA space to assign UMAP coordinates (credits to Changsheng Li for the implementation)}

\item{ortholog_table}{Dataframe for conversion between ortholog genes (by default package object \code{Hs2Mm.convert.table})}

\item{scGate_model}{scGate model used to filter target cell type from query data (if NULL use the model stored in \code{ref@misc$scGate})}

\item{ncores}{Number of cores for parallel execution (requires \code{BiocParallel})}
}
\value{
An augmented Seurat object with projected UMAP coordinates on the reference map and cell classifications
}
\description{
This function allows projecting single-cell RNA-seq datasets onto a reference map of cellular states. To project multiple dataset, submit a list of
Seurat objects with the query parameter.
}
\details{
See `load.reference.map()` to load or download a reference atlas.
}
\examples{
data(query_example_seurat)
ref <- load.reference.map()
make.projection(query_example_seurat, ref=ref)
}
