% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make.projection}
\alias{make.projection}
\title{Project a query scRNA-seq dataset onto a reference atlas}
\usage{
make.projection(
  query,
  ref = NULL,
  filter.cells = TRUE,
  scGate_model = NULL,
  query.assay = NULL,
  seurat.k.filter = 200,
  skip.normalize = FALSE,
  fast.mode = FALSE,
  ortholog_table = NULL,
  direct.projection = FALSE,
  ncores = 1,
  future.maxSize = 3000
)
}
\arguments{
\item{query}{Query data, either as single Seurat object or as a list of Seurat object}

\item{ref}{Reference Atlas - if NULL, downloads the default TIL reference atlas}

\item{filter.cells}{Pre-filter cells using `scGate`. Only set to FALSE if the dataset has been previously subset to desired cell type.}

\item{scGate_model}{scGate model used to filter target cell type from query data (if NULL use the model stored in \code{ref@misc$scGate})}

\item{query.assay}{Which assay slot to use for the query (defaults to DefaultAssay(query))}

\item{seurat.k.filter}{Integer. For alignment, how many neighbors (k) to use when picking anchors. Default is 200; try lower value in case of failure}

\item{skip.normalize}{By default, log-normalize the count data. If you have already normalized your data, you can skip normalization.}

\item{fast.mode}{Fast approximation for UMAP projection. Uses coordinates of nearest neighbors in PCA space to assign UMAP coordinates (credits to Changsheng Li for the implementation)}

\item{ortholog_table}{Dataframe for conversion between ortholog genes (by default package object \code{Hs2Mm.convert.table})}

\item{direct.projection}{If true, apply PCA transformation directly without alignment}

\item{ncores}{Number of cores for parallel execution (requires \code{future.apply})}

\item{future.maxSize}{For multi-core functionality, maximum allowed total size (in Mb) of global variables. To increment if required from \code{future.apply}}
}
\value{
An augmented Seurat object with projected UMAP coordinates on the reference map and cell classifications
}
\description{
This function allows projecting single-cell RNA-seq datasets onto a reference map of cellular states. To project multiple dataset, submit a list of
Seurat objects with the query parameter.
}
\details{
See `load.reference.map()` to load or download a reference atlas.
}
\examples{
data(query_example_seurat)
make.projection(query_example_seurat)
}
