% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{merge.Seurat.embeddings}
\alias{merge.Seurat.embeddings}
\title{Merge Seurat objects, including reductions (e.g. PCA, UMAP, ICA)}
\usage{
\method{merge}{Seurat.embeddings}(x = NULL, y = NULL, merge.dr = NULL, ...)
}
\arguments{
\item{x}{First object to merge}

\item{y}{Second object to merge}

\item{merge.dr}{Which reductions to merge. By default merges all reductions shared by the two objects.}

\item{...}{More parameters to \link{merge} function}
}
\value{
A merged Seurat object
}
\description{
Given two Seurat objects, merge counts and data as well as dim reductions (PCA, UMAP, ICA, etc.)
}
\examples{
seurat.merged <- merge.Seurat.embeddings(obj.1, obj.2)  
#To merge multiple object stored in a list
seurat.merged <- Reduce(f=merge.Seurat.embeddings, x=obj.list)
}
