% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.projection}
\alias{plot.projection}
\title{Show UMAP projection of query on reference map}
\usage{
\method{plot}{projection}(
  ref,
  query = NULL,
  labels.col = "functional.cluster",
  cols = NULL,
  linesize = 1,
  pointsize = 1,
  ref.alpha = 0.3,
  ref.size = NULL,
  ...
)
}
\arguments{
\item{ref}{Reference Atlas}

\item{query}{Seurat object with query data}

\item{labels.col}{The metadata field to annotate the clusters (default: functional.cluster)}

\item{cols}{Custom color palette for clusters}

\item{linesize}{Contour line thickness for projected query}

\item{pointsize}{Point size for cells in projected query}

\item{ref.alpha}{Transparency parameter for reference cells}

\item{ref.size}{Adjust point size for reference cells}

\item{...}{Additional parameters for \code{DimPlot}, e.g. raster=T to
limit image size}
}
\value{
UMAP plot of reference map with projected query set in the same space
}
\description{
Plots the UMAP representation of the reference map, together with the projected coordinates of a query dataset.
}
\examples{
data(query_example_seurat)
ref <- load.reference.map()
q <- Run.ProjecTILs(query_example_seurat, ref=ref, fast.umap.predict=TRUE)
plot.projection(ref=ref, query=q)
}
