% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.statepred.composition}
\alias{plot.statepred.composition}
\title{Summarize the predicted cell states of an object}
\usage{
\method{plot}{statepred.composition}(
  ref,
  query,
  labels.col = "functional.cluster",
  cols = NULL,
  metric = c("Count", "Percent")
)
}
\arguments{
\item{ref}{Seurat object with the reference object}

\item{query}{Seurat object with query data}

\item{labels.col}{The metadata field used to annotate the clusters (default: functional.cluster)}

\item{cols}{Custom color palette for clusters}

\item{metric}{One of `Count` or `Percent`. `Count` plots the absolute number of cells, `Percent` the fraction on the total number of cells.}
}
\value{
Barplot of predicted state composition
}
\description{
Makes a barplot of the frequency of cell states in a query object.
}
\examples{
plot.statepred.composition(query_example.seurat)
}
