% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.states.radar}
\alias{plot.states.radar}
\title{Show expression level of key genes}
\usage{
plot.states.radar(ref)
}
\arguments{
\item{ref}{Reference Atlas}

\item{query}{Query data, either as a Seurat object or as a list of Seurat objects}

\item{labels.col}{The metadata field used to annotate the clusters}

\item{ref.assay}{The assay to pull the reference expression data}

\item{query.assay}{The assay to pull the query expression data}

\item{genes4radar}{Which genes to use for plotting}

\item{meta4radar}{Which metadata columns (numeric) to use for plotting. If not NULL, \code{genes4radar} are ignored}

\item{min.cells}{Only display cell states with a minimum number of cells}

\item{cols}{Custom color palette for samples in radar plot}

\item{return}{Return the combined plots instead of printing them to the default device (deprecated)}

\item{return.as.list}{Return plots in a list, instead of combining them in a single plot}
}
\value{
Radar plot of gene expression of key genes by cell subtype
}
\description{
Makes a radar plot of the expression level of a set of genes. It can be useful to compare
the gene expression profile of different cell states in the reference atlas vs. a projected set.
}
\examples{
ref <- load.reference.map()
plot.states.radar(ref)
}
